INS-2760: move pubsubwrap package to instrumentation/introspector/pubsubwrap
