+++ instrumentation/introspector/pubsubwrap/middleware_stat.go
