+++ server/internal/heavy/components.go
@@ -212,7 +212,6 @@ func newComponents(ctx context.Context, cfg configuration.Configuration, genesis
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to start MessageBus")
 		}
-		// TODO: only use wrapped pubSub here?
 		WmBus = bus.NewBus(pubSub, Pulses, Coordinator, CryptoScheme)
 	}
